Rewrite the github automaiton proj and adjust it to work with the new seperate infra project
