+++ GitHubAutomation/Pages/LoginPage.cs
@@ -1,6 +1,6 @@
 ﻿using OpenQA.Selenium;
 
-namespace GitHub.pages.GitHubPages
+namespace GitHubAutomation.Pages
 {
     internal class LoginPage : GitHubPage
     {
@@ -8,20 +8,20 @@ namespace GitHub.pages.GitHubPages
         private readonly By passwordField;
         private readonly By signInBtn;
 
-        public LoginPage ()
+        public LoginPage(IWebDriver driver, ILogger logger) : base(driver, logger)
         {
             userNameField = By.Id("login_field");
             passwordField = By.Id("password");
             signInBtn = By.CssSelector("#login [type='submit']");
         }
-        public override bool IsDisplayed() => IsElementVisible(signInBtn);
+        public override bool IsDisplayed() => Bot.IsElementVisible(signInBtn);
 
         public UserHomePage Signin(string userName, string password)
         {
-            SendKeys(userNameField, userName, nameof(userNameField));
-            SendKeys(passwordField, password, nameof(passwordField));
-            Click(signInBtn, nameof(signInBtn));
-            return new UserHomePage();
+            Bot.SendKeys(userNameField, userName, nameof(userNameField));
+            Bot.SendKeys(passwordField, password, nameof(passwordField));
+            Bot.Click(signInBtn, nameof(signInBtn));
+            return new UserHomePage(driver, logger);
         }
     }
 }
