+++ GitHubAutomation/Pages/NewRepPage.cs
@@ -1,6 +1,6 @@
 ﻿using OpenQA.Selenium;
 
-namespace GitHub.pages.GitHubPages
+namespace GitHubAutomation.Pages
 {
     internal class NewRepPage : GitHubPage
     {
@@ -9,32 +9,32 @@ namespace GitHub.pages.GitHubPages
         private readonly By repDescriptionField;
         private readonly By createRepBtn;
 
-        public NewRepPage()
+        public NewRepPage(IWebDriver driver, ILogger logger) : base(driver, logger)
         {
-            repDescriptionField = By.Name("Description");
             repNameField = By.Id("react-aria-2");
             validationText = By.Id("react-aria-2-validationMessage");
+            repDescriptionField = By.Name("Description");
             createRepBtn = By.XPath("//button//*[contains(text(),'Create repository')]");
         }
 
-        public override bool IsDisplayed() => IsElementVisible(repNameField);
+        public override bool IsDisplayed() => Bot.IsElementVisible(repNameField);
 
         public RepHomePage CreateNewRep(string repName, string repDescription)
         {
-            SendKeys(repNameField, repName, nameof(repNameField));
+            Bot.SendKeys(repNameField, repName, nameof(repNameField));
             IsRepNameValid();
 
-            SendKeys(repDescriptionField, repDescription, nameof(repDescriptionField));
-
-            ScrollToElement(createRepBtn);
-            Click(createRepBtn, nameof(createRepBtn));
-            return new RepHomePage();
+            Bot.SendKeys(repDescriptionField, repDescription, nameof(repDescriptionField));
+            
+            Bot.ScrollToElement(createRepBtn);
+            Bot.Click(createRepBtn, nameof(createRepBtn));
+            return new RepHomePage(driver, logger);
         }
 
-        public void IsRepNameValid() {
-
+        private void IsRepNameValid()
+        {
             string[] validTexts = { "is available", "repository will be created as" };
-            WaitForElementToContainAnyText(validationText, validTexts);
+            Bot.WaitForElementToContainAnyText(validationText, validTexts);
         }
     }
 }
