--- GitHubProjectFolder/GitHub.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-
-	<!--TODO: if Driver Manager doesn't working try to uncomment this line-->
-	<!--<RuntimeIdentifier>win-x64</RuntimeIdentifier>-->
-
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="DotNetSeleniumExtras.WaitHelpers" Version="3.11.0" />
-    <PackageReference Include="NUnit" Version="3.13.3" />
-    <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.2.0" />
-    <PackageReference Include="Selenium.Support" Version="4.2.0" />
-    <PackageReference Include="Selenium.WebDriver" Version="4.2.0" />
-    <PackageReference Include="Selenium.WebDriver.ChromeDriver" Version="113.0.5672.6300" />
-    <PackageReference Include="WebDriverManager" Version="2.16.2" />
-  </ItemGroup>
-</Project>
