--- GitHubProjectFolder/config/settingsFile.RunSettings
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
-<RunSettings>
-	<!-- Parameters used by tests at runtime -->
-	<TestRunParameters>
-		<Parameter name="local" value="true" />	<!--true/false-->
-	</TestRunParameters>
-</RunSettings>
