--- GitHubProjectFolder/pages/GitHubPages/GitHubPage.cs
@@ -1,9 +1,0 @@
-﻿using GitHub.config;
-
-namespace GitHub.pages.GitHubPages
-{
-    internal abstract class GitHubPage : BasePage
-    {
-        public abstract bool IsDisplayed();
-    }
-}
